pyproject.toml
[project]
name = "redmine-issue-bulk-creator"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Kazuki Arai", email = "zukain5@gmail.com" }
]
dependencies = [
    "python-redmine>=2.4.0",
    "questionary>=2.0.1",
    "python-dotenv>=1.0.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ipython>=8.12.3",
]

[tool.hatch.metadata]
allow-direct-references = true
